integracion firebase en cotizaciones
se integró base de datos de firebase para el registro de cotizaciones y se organizó el modulo completo de cotizaciones de forma modularizada separando logica y ui (carpetas dto, mappers, datasource,etc)

+++ lib/features/cotizacion/presentation/widgets/mano_obra_dialog.dart
@@ -264,7 +264,7 @@ class _ManoObraDialogState extends State<ManoObraDialog> {
             final item = ManoDeObra(
               cargo: cargoSeleccionado,
               valorJornada: double.parse(valorJornadaController.text),
-              dias: double.parse(diasController.text),
+              dias: int.parse(diasController.text),
             );
 
             Navigator.pop(context, item);
